requirements.txt
langchain[google-genai]
langchain-groq
langgraph
yfinance
pandas
ta
duckduckgo-search
ddgs
python-dotenv
structlog
